+++ source_bundles/__init__.py
@@ -1,0 +1,2 @@
+"""Product source bundle registry helpers."""
+